[Alice/Amelyn/Para] Addressed some comments for PR 183
