+++ backend/package.json
@@ -8,7 +8,7 @@
     "ci": "yarn lint && node bin/test-seed.js && yarn test",
     "dev": "yarn nodemon -L index.js",
     "lint": "yarn eslint src spec",
-    "e2e-seed": "NODE_ENV=test node bin/e2e-seed-new.js",
+    "e2e-seed": "NODE_ENV=test node bin/e2e-seed.js",
     "test": "NODE_ENV=test yarn mocha --reporter dot --check-leaks --timeout 10000 -r spec/specGlobals --recursive spec",
     "tdd": "NODE_ENV=test yarn mocha --reporter dot --check-leaks --timeout 10000 -r spec/specGlobals --recursive spec/unit --watch",
     "seed": "node bin/seed.js && node bin/test-seed.js",
