--- e2e/cypress/fixtures/profile.json
@@ -1,5 +1,0 @@
-{
-  "id": 8739,
-  "name": "Jane",
-  "email": "jane@example.com"
-}
