+++ go.sh
@@ -2,14 +2,6 @@
 
 set -e
 
-# This follows a different path to the other tests/shell, since the e2e script
-# spins up its own containers as needed.
-if [ $1 = "e2e" ]
-then
-  ./bin/e2e.sh
-  exit
-fi
-
 # Usage examples:
 # `./go.sh` to get an interactive shell in a container for development
 # `./go.sh yarn test` to run the tests in a container and then exit
@@ -17,7 +9,7 @@ docker-compose up -d --remove-orphans
 
 if [[ -z $1 ]];
 then
-  docker-compose exec core /bin/bash
+  docker-compose exec core /bin/sh
 else
   docker-compose exec core "$@"
 fi
